Refactoring the single microcoapy file into multiple otheres to optimize memory usage  - This change as necessary for the support of ESP8266 that could not load the single file due to memory error.  - The library can be used by importing the folder instead of specific files.
