+++ microcoapy/__init__.py
@@ -1,0 +1,3 @@
+from microcoapy.microcoapy import Coap
+from microcoapy.coap_macros import COAP_CONTENT_TYPE
+from microcoapy.coap_macros import COAP_RESPONSE_CODE
